FIX: Custom Decimal attribute

+++ database/migrations/2019_01_09_031229_create_customers_table.php
@@ -32,9 +32,9 @@ class CreateCustomersTable extends Migration
 
             $table->boolean('with_tax')->default(0);
             $table->boolean('with_pph')->default(0);
-            $table->float('tax')->default(0);
-            $table->float('pph_service')->default(0);
-            $table->float('pph_material')->default(0);
+            $table->double('tax')->default(0);
+            $table->double('pph_service')->default(0);
+            $table->double('pph_material')->default(0);
 
             $table->string('invoice_mode')->nullable();
             $table->string('delivery_mode')->nullable();
